+++ src/base-lib/src/common/utils.ts
@@ -12,9 +12,10 @@
  */
 
 import { randomInt } from 'crypto';
+import moment from 'moment-timezone';
 import Mustache from 'mustache';
-import { logger } from './powertools';
 import { MUSTACHE_RENDER_CATEGORIES } from '../constant';
+import { logger } from './powertools';
 
 export function isEmpty(a: any): boolean {
   if (a === '') return true; //Verify empty string
@@ -107,3 +108,16 @@ export function parseMetadataFromSql(fileContent: string): any[] {
   }
   return metadataArray;
 }
+
+export const getDateTimeWithTimezoneString = (
+  timestamp: number,
+  timezone: string,
+) => {
+  return moment
+    .tz(moment(timestamp), timezone)
+    .format('YYYY-MM-DD HH:mm:ss(z)');
+};
+
+export const getLocaleDateString = (timestamp: number, timezone: string) => {
+  return moment.tz(moment(timestamp), timezone).format('YYYY-MM-DD');
+};
